[CINN] Fix name duplication for TileTransposeTactic
